+++ sys/posix/gen_calls.c
@@ -1081,6 +1081,8 @@ int sys_uname(struct thread *td, struct sys_uname_args *args)
 	strncpy(uts->version, UBIXOS_VERSION_STRING, sizeof(uts->version) - 1);
 #if defined(__aarch64__)
 	strncpy(uts->machine, "aarch64", sizeof(uts->machine) - 1);
+#elif defined(__x86_64__)
+	strncpy(uts->machine, "x86_64", sizeof(uts->machine) - 1);
 #else
 	strncpy(uts->machine, "i386", sizeof(uts->machine) - 1);
 #endif